Move event processing from Node to BaseNode
