+++ cocos/core/scene-graph/node-event-processor.ts
@@ -35,10 +35,10 @@ import { EventListener } from '../platform/event-manager/event-listener';
 import { eventManager } from '../platform/event-manager/event-manager';
 import { EventMouse, EventTouch } from '../platform/event-manager/events';
 import { Touch } from '../platform/event-manager/touch';
-import { INode } from '../utils/interfaces';
+import { IBaseNode, INode } from '../utils/interfaces';
 
-const _cachedArray = new Array<INode>(16);
-let _currentHovered: INode | null = null;
+const _cachedArray = new Array<IBaseNode>(16);
+let _currentHovered: IBaseNode | null = null;
 let pos = new Vec2();
 
 const _touchEvents = [
@@ -59,7 +59,7 @@ const _mouseEvents = [
 
 // TODO: rearrange event
 function _touchStartHandler (this: EventListener, touch: Touch, event: EventTouch) {
-    const node = this.owner as INode;
+    const node = this.owner as IBaseNode;
     // @ts-ignore
     if (!node || !node.uiTransfromComp) {
         return false;
@@ -213,8 +213,8 @@ function _mouseWheelHandler (this: EventListener, event: EventMouse) {
     }
 }
 
-function _doDispatchEvent (owner: INode, event: Event) {
-    let target: INode;
+function _doDispatchEvent (owner: IBaseNode, event: Event) {
+    let target: IBaseNode;
     let i = 0;
     event.target = owner;
 
@@ -287,7 +287,7 @@ function _searchMaskInParent (node: INode | null) {
     return null;
 }
 
-function _checkListeners (node: INode, events: string[]) {
+function _checkListeners (node: IBaseNode, events: string[]) {
     if (!node._persistNode) {
         let i = 0;
         if (node.eventProcessor.bubblingTargets) {
@@ -314,7 +314,7 @@ function _checkListeners (node: INode, events: string[]) {
  * 节点事件类。
  */
 export class NodeEventProcessor {
-    public get node (): INode {
+    public get node (): IBaseNode {
         return this._node;
     }
 
@@ -340,21 +340,21 @@ export class NodeEventProcessor {
      */
     public mouseListener: EventListener | null = null;
 
-    private _node: INode;
+    private _node: IBaseNode;
 
-    constructor (node: INode) {
+    constructor (node: IBaseNode) {
         this._node = node;
     }
 
     public reattach (): void {
         if (this.touchListener) {
-            const mask = this.touchListener.mask = _searchMaskInParent(this._node);
+            const mask = this.touchListener.mask = _searchMaskInParent(this._node as INode);
             if (this.mouseListener) {
                 this.mouseListener.mask = mask;
             }
         }
         else if (this.mouseListener) {
-            this.mouseListener.mask = _searchMaskInParent(this._node);
+            this.mouseListener.mask = _searchMaskInParent(this._node as INode);
         }
     }
 
@@ -615,7 +615,7 @@ export class NodeEventProcessor {
      * @param type - 一个监听事件类型的字符串。
      * @param array - 接收目标的数组。
      */
-    public getCapturingTargets (type: string, targets: INode[]) {
+    public getCapturingTargets (type: string, targets: IBaseNode[]) {
         let parent = this._node.parent;
         while (parent) {
             if (parent.eventProcessor.capturingTargets && parent.eventProcessor.capturingTargets.hasEventListener(type)) {
@@ -634,7 +634,7 @@ export class NodeEventProcessor {
      * @param type - 一个监听事件类型的字符串。
      * @param array - 接收目标的数组。
      */
-    public getBubblingTargets (type: string, targets: INode[]) {
+    public getBubblingTargets (type: string, targets: IBaseNode[]) {
         let parent = this._node.parent;
         while (parent) {
             if (parent.eventProcessor.bubblingTargets && parent.eventProcessor.bubblingTargets.hasEventListener(type)) {
@@ -656,7 +656,7 @@ export class NodeEventProcessor {
                     event: cc.EventListener.TOUCH_ONE_BY_ONE,
                     swallowTouches: true,
                     owner: this._node,
-                    mask: _searchMaskInParent(this._node),
+                    mask: _searchMaskInParent(this._node as INode),
                     onTouchBegan: _touchStartHandler,
                     onTouchMoved: _touchMoveHandler,
                     onTouchEnded: _touchEndHandler,
@@ -672,7 +672,7 @@ export class NodeEventProcessor {
                     event: cc.EventListener.MOUSE,
                     _previousIn: false,
                     owner: this._node,
-                    mask: _searchMaskInParent(this._node),
+                    mask: _searchMaskInParent(this._node as INode),
                     onMouseDown: _mouseDownHandler,
                     onMouseMove: _mouseMoveHandler,
                     onMouseUp: _mouseUpHandler,
