+++ cocos/core/scene-graph/node.ts
@@ -29,21 +29,18 @@
 
 import { UIComponent, UITransformComponent } from '../components/ui-base';
 import { ccclass, property } from '../data/class-decorator';
-import Event from '../event/event';
 import { Mat3, Mat4, Quat, Size, Vec2, Vec3 } from '../math';
 import { SystemEventType } from '../platform/event-manager/event-enum';
 import { eventManager } from '../platform/event-manager/event-manager';
 import { INode } from '../utils/interfaces';
-import { BaseNode } from './base-node';
+import { BaseNode, TRANFORM_ON } from './base-node';
 import { Layers } from './layers';
 import { NodeSpace, TransformDirtyBit } from './node-enum';
-import { NodeEventProcessor } from './node-event-processor';
 
 const v3_a = new Vec3();
 const q_a = new Quat();
 const q_b = new Quat();
 const array_a = new Array(10);
-const TRANFORM_ON = 1 << 0;
 const qt_1 = new Quat();
 const m3_1 = new Mat3();
 const m3_scaling = new Mat3();
@@ -109,8 +106,6 @@ export class Node extends BaseNode implements INode {
     protected _dirtyFlags = TransformDirtyBit.NONE; // does the world transform need to update?
     protected _eulerDirty = false;
 
-    protected _eventProcessor: NodeEventProcessor = new NodeEventProcessor(this as INode);
-    protected _eventMask = 0;
     private _uiTransfromComp: UITransformComponent | null = null;
     // tslint:disable-next-line: member-ordering
     public _uiComp: UIComponent | null = null;
@@ -311,10 +306,6 @@ export class Node extends BaseNode implements INode {
     }
     set anchorY (value: number) {
         this.uiTransfromComp!.anchorY = value;
-    }
-
-    get eventProcessor () {
-        return this._eventProcessor;
     }
 
     // ===============================
@@ -946,55 +937,6 @@ export class Node extends BaseNode implements INode {
 
     public setContentSize (size: Size | number, height?: number) {
         this.uiTransfromComp!.setContentSize(size, height);
-    }
-
-    // Event: maybe remove
-
-    public on (type: string | SystemEventType, callback: Function, target?: Object, useCapture?: any) {
-        switch (type) {
-            case SystemEventType.TRANSFORM_CHANGED:
-                this._eventMask |= TRANFORM_ON;
-                break;
-        }
-        this._eventProcessor.on(type, callback, target, useCapture);
-    }
-
-    public off (type: string, callback?: Function, target?: Object, useCapture?: any) {
-        this._eventProcessor.off(type, callback, target, useCapture);
-
-        const hasListeners = this._eventProcessor.hasEventListener(type);
-        // All listener removed
-        if (!hasListeners) {
-            switch (type) {
-                case SystemEventType.TRANSFORM_CHANGED:
-                    this._eventMask &= ~TRANFORM_ON;
-                    break;
-            }
-        }
-    }
-
-    public once (type: string, callback: Function, target?: Object, useCapture?: any) {
-        this._eventProcessor.once(type, callback, target, useCapture);
-    }
-
-    public emit (type: string, ...args: any[]) {
-        this._eventProcessor.emit(type, ...args);
-    }
-
-    public dispatchEvent (event: Event) {
-        this._eventProcessor.dispatchEvent(event);
-    }
-
-    public hasEventListener (type: string) {
-        return this._eventProcessor.hasEventListener(type);
-    }
-
-    public targetOff (target: string | Object) {
-        this._eventProcessor.targetOff(target);
-        // Check for event mask reset
-        if ((this._eventMask & TRANFORM_ON) && !this._eventProcessor.hasEventListener(SystemEventType.TRANSFORM_CHANGED)) {
-            this._eventMask &= ~TRANFORM_ON;
-        }
     }
 
     public pauseSystemEvents (recursive: boolean): void {
